Created prototype script for apriltags
Created prototype script for apriltags so we can begin getting them to work for both auto mode as well as to tune tracking so it works for our needs. Also renamed the other robot prototype script, enabled some of the sample scripts for testing, and fixed a error in @ColoradoBiker script.